fd support for mmap(2)
we already had support to mmap files for fast exec(2), so most of the work was
already done. had to add new error paths since there can now be pagefaults that
cannot be handled when a mapped file is truncated before its first access
through the mapping.

+++ biscuit/user/c/include/litc.h
@@ -65,7 +65,7 @@ extern "C" {
 #define		EINPROGRESS	115
 #define		ERRNO_LAST	115
 
-#define		MAP_FAILED	((void *) -1)
+#define		MAP_FAILED	((void *)(long) -1)
 
 #define		MAP_SHARED	0x01
 #define		MAP_PRIVATE	0x02
